Refactored methods from Runner into BaseRunner
